Moved sample to its own package, fixed class visibility
