+++ sample/src/main/java/com/timehop/stickyheadersrecyclerview/sample/RecyclerArrayAdapter.java
@@ -1,4 +1,4 @@
-package com.timehop.stickyheadersrecyclerview;
+package com.timehop.stickyheadersrecyclerview.sample;
 
 import android.support.v7.widget.RecyclerView;
